Adopt C++20 and format game code

+++ src/game/audio.cpp
@@ -7,7 +7,14 @@ namespace
 {
     constexpr int SAMPLE_RATE = 44100;
     Sound sounds[static_cast<int>(SfxType::COUNT)];
-    Sound GenerateSfx(float duration, float freq_start, float freq_end, float amp, int wave_type)
+    Sound GenerateSfx
+    (
+        float duration,
+        float freq_start,
+        float freq_end,
+        float amp,
+        int wave_type
+    )
     {
         int n = static_cast<int>(SAMPLE_RATE * duration);
 
@@ -33,7 +40,7 @@ namespace
             }
             else if (wave_type == 2)
             {
-                val = static_cast<float>(rand()) / RAND_MAX * 2.0f - 1.0f;
+                val = static_cast<float>(rand()) / static_cast<float>(RAND_MAX) * 2.0f - 1.0f;
             }
             else
             {
@@ -50,7 +57,7 @@ namespace
 
     Sound GenerateArpeggio()
     {
-        float notes[] = { 400.0f, 600.0f, 900.0f };
+        float notes[] = {400.0f, 600.0f, 900.0f};
         float note_dur = 0.08f;
         float gap = 0.02f;
         int note_samples = static_cast<int>((SAMPLE_RATE * note_dur));
@@ -91,13 +98,20 @@ namespace
 void InitAudio()
 {
     InitAudioDevice();
-    sounds[static_cast<int>(SfxType::PLACE_GATE)]     = GenerateSfx(0.06f, 600, 200, 1.0f, 0);
-    sounds[static_cast<int>(SfxType::REMOVE_GATE)]    = GenerateSfx(0.05f, 100, 100, 0.4f, 2);
-    sounds[static_cast<int>(SfxType::CONNECT_WIRE)]   = GenerateSfx(0.05f, 800, 800, 0.8f, 1);
-    sounds[static_cast<int>(SfxType::DISCONNECT_WIRE)] = GenerateSfx(0.04f, 600, 200, 0.6f, 0);
-    sounds[static_cast<int>(SfxType::TOGGLE_INPUT)]   = GenerateSfx(0.03f, 1000, 1000, 0.7f, 0);
-    sounds[static_cast<int>(SfxType::SOLVED)]         = GenerateArpeggio();
-    sounds[static_cast<int>(SfxType::ERROR)]          = GenerateSfx(0.08f, 150, 150, 0.5f, 1);
+    sounds[static_cast<int>(SfxType::PLACE_GATE)] =
+        GenerateSfx(0.06f, 600, 200, 1.0f, 0);
+    sounds[static_cast<int>(SfxType::REMOVE_GATE)] =
+        GenerateSfx(0.05f, 100, 100, 0.4f, 2);
+    sounds[static_cast<int>(SfxType::CONNECT_WIRE)] =
+        GenerateSfx(0.05f, 800, 800, 0.8f, 1);
+    sounds[static_cast<int>(SfxType::DISCONNECT_WIRE)] =
+        GenerateSfx(0.04f, 600, 200, 0.6f, 0);
+    sounds[static_cast<int>(SfxType::TOGGLE_INPUT)] =
+        GenerateSfx(0.03f, 1000, 1000, 0.7f, 0);
+    sounds[static_cast<int>(SfxType::SOLVED)] =
+        GenerateArpeggio();
+    sounds[static_cast<int>(SfxType::ERROR)] =
+        GenerateSfx(0.08f, 150, 150, 0.5f, 1);
 }
 
 void PlaySfx(SfxType type)
@@ -107,6 +121,6 @@ void PlaySfx(SfxType type)
 
 void ShutdownAudio()
 {
-    for (int i = 0; i < static_cast<int>(SfxType::COUNT); i++) UnloadSound(sounds[i]);
+    for (int i = 0; i < static_cast<int>(SfxType::COUNT); i++)  UnloadSound(sounds[i]);
     CloseAudioDevice();
 }
